move under ca_custom_root test folder
